+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Create script with event(click) handlers using `data-*` attributes
+
 ## [v2.5.4] - 2020-12-01
 
 ### Added